Updated the input to reflect controls.md
The preset saving feature still does not work. One problem is that when a preset is copied, it is of a different form than the original copied bay. This is because when you add in objects to a copied preset, creating a preset of that bay causes it to lose all of its children.